Move all ProbeCommands to the Probe namespace
