+++ src/Kernel.php
@@ -3,7 +3,7 @@
 namespace App;
 
 use App\DependencyInjection\Compiler\CommandPass;
-use App\ShellCommand\CommandInterface;
+use App\Probe\CommandInterface;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\Config\Resource\FileResource;
