+++ src/Probe/CommandInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\ShellCommand;
+namespace App\Probe;
 
 interface CommandInterface
 {
