+++ src/Probe/GetConfiguration.php
@@ -1,14 +1,14 @@
 <?php
 declare(strict_types=1);
 
-namespace App\ShellCommand;
+namespace App\Probe;
 
 use App\Client\FirepingClient;
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Psr7\Request;
 use Psr\Log\LoggerInterface;
 
-class GetConfigHttpWorkerCommand implements CommandInterface
+class GetConfiguration implements CommandInterface
 {
     /**
      * @var LoggerInterface
